+++ navigation/BottomTabNavigator.tsx
@@ -1,0 +1,64 @@
+import { FontAwesome } from '@expo/vector-icons';
+import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
+import { ColorSchemeName, Pressable } from 'react-native';
+
+import SearchStackNavigator from './SearchStackNavigator';
+import BoardStackNavigator from './BoardStackNavigator';
+import ProfileStackNavigator from './ProfileStackNavigator';
+
+import { RootTabParamList, RootTabScreenProps } from '../types';
+
+const Tab = createBottomTabNavigator<RootTabParamList>();
+
+export default function BottomTabNavigator() {
+  return (
+    <Tab.Navigator
+      initialRouteName="TabOne"
+      screenOptions={{
+        tabBarActiveTintColor: '#2f95dc',
+      }}
+    >
+      <Tab.Screen name="Search" component={SearchStackNavigator} />
+      <Tab.Screen name="Board" component={BoardStackNavigator} />
+      <Tab.Screen name="Profile" component={ProfileStackNavigator} />
+      {/* <Tab.Screen
+        name="TabOne"
+        component={TabOneScreen}
+        options={({ navigation }: RootTabScreenProps<'TabOne'>) => ({
+          title: 'Tab One',
+          tabBarIcon: ({ color }) => <TabBarIcon name="code" color={color} />,
+          headerRight: () => (
+            <Pressable
+              onPress={() => navigation.navigate('Modal')}
+              style={({ pressed }) => ({
+                opacity: pressed ? 0.5 : 1,
+              })}
+            >
+              <FontAwesome
+                name="info-circle"
+                size={25}
+                color={'#000'}
+                style={{ marginRight: 15 }}
+              />
+            </Pressable>
+          ),
+        })}
+      />
+      <Tab.Screen
+        name="TabTwo"
+        component={TabTwoScreen}
+        options={{
+          title: 'Tab Two',
+          tabBarIcon: ({ color }) => <TabBarIcon name="code" color={color} />,
+        }}
+      /> */}
+    </Tab.Navigator>
+  );
+}
+
+function TabBarIcon(props: {
+  name: React.ComponentProps<typeof FontAwesome>['name'];
+  color: string;
+}) {
+  return <FontAwesome size={30} style={{ marginBottom: -3 }} {...props} />;
+}
