+++ navigation/SearchStackNavigator.tsx
@@ -1,0 +1,20 @@
+import { createNativeStackNavigator } from '@react-navigation/native-stack';
+import SearchScreen from '../screens/SearchScreen';
+
+import { RootStackParamList } from '../types';
+
+const Stack = createNativeStackNavigator<RootStackParamList>();
+
+export default function SearchStackNavigator() {
+  return (
+    <Stack.Navigator>
+      <Stack.Screen
+        name="SearchStack"
+        component={SearchScreen}
+        options={{
+          headerShown: false,
+        }}
+      />
+    </Stack.Navigator>
+  );
+}
