--- screens/TabOneScreen.tsx
@@ -1,32 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-import { Text, View } from 'react-native';
-import { RootTabScreenProps } from '../types';
-
-export default function TabOneScreen({
-  navigation,
-}: RootTabScreenProps<'TabOne'>) {
-  return (
-    <View style={styles.container}>
-      <Text style={styles.title}>Tab One</Text>
-      <View style={styles.separator} />
-    </View>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  title: {
-    fontSize: 20,
-    fontWeight: 'bold',
-  },
-  separator: {
-    marginVertical: 30,
-    height: 1,
-    width: '80%',
-  },
-});
